fixed rpmlint warnings

+++ courier-imap-x/libs/maildir/maildirwatch.c
@@ -28,7 +28,8 @@ static void alarm_handler(int signum)
 	static const char msg[]=
 		"Timeout initializing the FAM library. Your FAM library is broken.\n";
 
-	(void)write(2, msg, sizeof(msg)-1);
+	if (write(2, msg, sizeof(msg)-1) == -1)
+		;
 	kill(getpid(), SIGKILL);
 }
 #endif
